Move settings to the store, completely
