+++ src/tc-renderer/store/modules/helpers/settings/default-settings.js
