+++ src/tc-renderer/store/modules/helpers/settings/settings-file-path.js
@@ -1,0 +1,7 @@
+import path from 'path'
+import electron from 'electron'
+
+export const settingsFilePath = path.resolve(
+  electron.remote.app.getPath('userData'),
+  'settings.json'
+)
