+++ src/tc-renderer/store/modules/helpers/settings/validate-settings.js
